Add shuffle of files in transaction package
