+++ packages/transactions/src/OutboundShipment/index.ts
@@ -1,0 +1,2 @@
+export { OutboundShipmentListView } from './ListView';
+export { OutboundShipmentDetailView } from './DetailView';
